+++ AoC/2019/Classes/IntcodeProgramInput.cs
@@ -4,14 +4,14 @@ namespace AoC._2019.Classes
 {
     public class IntcodeProgramInput
     {
-        public Action<int[]> MemoryInitialisation { get; }
-        public int[] Inputs { get; }
+        public Action<long[]> MemoryInitialisation { get; }
+        public long[] Inputs { get; }
 
-        public IntcodeProgramInput(Action<int[]> memoryInitialisation = null, int[] inputs = null)
+        public IntcodeProgramInput(Action<long[]> memoryInitialisation = null, long[] inputs = null)
         {
-            MemoryInitialisation = memoryInitialisation ?? new Action<int[]>(mem => { return; });
+            MemoryInitialisation = memoryInitialisation ?? new Action<long[]>(mem => { return; });
 
-            Inputs = new int[inputs == null ? 0 : inputs.Length];
+            Inputs = new long[inputs == null ? 0 : inputs.Length];
             if (inputs != null && inputs.Length > 0) inputs.CopyTo(Inputs, 0);
         }
     }
