+++ AoC/2019/Classes/OperationInput.cs
@@ -4,10 +4,10 @@ namespace AoC._2019.Classes
 {
     public class OperationInput
     {
-        public Func<int> GetInput { get; }
-        public int[] Parameters { get; }
+        public Func<long> GetInput { get; }
+        public long[] Parameters { get; }
 
-        public OperationInput(Func<int> getInput, params int[] parameters)
+        public OperationInput(Func<long> getInput, params long[] parameters)
         {
             GetInput = getInput;
             Parameters = parameters;
